extension: extract some hooks
